Add some helper types
Some constants and types are given by the test so they will not change
during the development but let's pretend we want to be able to easily
update some types later on. In that case it would be nice to not have to
change every function using those types but to have one central place
where types and constants are managed for the whole project.