Add Transaction enum
That enum will contain all transactions we want to support.
Using an enum was preferred over a single struct modeling Transactions
because it allows better type safety (e.g. forbid Disputes which hold an
amount).

Because this exercise requires us to create Transactions from csv rows,
a function which takes a csv::StringRecord to construct a Transaction is
very ergonomic.