Test creating Transaction from CSV rows
This function takes unsanitized and potentially malicious user data and
creates objects which the rest of the code base trusts. We need to make
sure that the parsing is correct and no malformed can be used to
generate valid Transactions.

Each supported transaction types gets one positive and a couple negative
tests.