Implement TransactionRecord
A TransactionRecord keeps track of the DiputeState for each executed
Transaction. The impl block should enforce the state machine of
disputing a Transaction as is described in the exercise.

Ideally we would only utilize the Rust type system to implement a state
machine. This could be achived by implemnting state transition with
functions on states like this:
struct DisputedState;
impl for DisputedState {
    pub fn resolve(self) -> ResolvedState;
    pub fn chargeback(self) -> ChargedBackState;
}
Unfortunately I was not able to implement that ergonomically while still
being able to store that state in a collection.

The tradeoff was using an enum while asserting correct enum variants on
the functions implementing the state transition.